split usage file
